reorg some defns and format constants
