+++ src/bellatrix/presets/mainnet.rs
@@ -10,11 +10,11 @@ use crate::phase0::mainnet::{
 
 pub use bellatrix::*;
 
-pub const INACTIVITY_PENALTY_QUOTIENT_BELLATRIX: u64 = 16777216;
+pub const INACTIVITY_PENALTY_QUOTIENT_BELLATRIX: u64 = 16_777_216;
 pub const MIN_SLASHING_PENALTY_QUOTIENT_BELLATRIX: u64 = 32;
 pub const PROPORTIONAL_SLASHING_MULTIPLIER_BELLATRIX: u64 = 3;
-pub const MAX_BYTES_PER_TRANSACTION: usize = 1073741824;
-pub const MAX_TRANSACTIONS_PER_PAYLOAD: usize = 1048576;
+pub const MAX_BYTES_PER_TRANSACTION: usize = 1_073_741_824;
+pub const MAX_TRANSACTIONS_PER_PAYLOAD: usize = 1_048_576;
 pub const BYTES_PER_LOGS_BLOOM: usize = 256;
 pub const MAX_EXTRA_DATA_BYTES: usize = 32;
